Full migration for cpu_anim, migration draft for bevy_game

+++ bevy_game/src/level/assets/tileset_indexing.rs
@@ -2,13 +2,13 @@ use std::collections::HashMap;
 
 pub enum TilesetIndexing {
     Continious,
-    Special(HashMap<u32, u16>),
+    Special(HashMap<u32, u32>),
 }
 
 impl TilesetIndexing {
-    pub fn dispatch(&self, x: u32) -> u16 {
+    pub fn dispatch(&self, x: u32) -> u32 {
         match self {
-            TilesetIndexing::Continious => x as u16,
+            TilesetIndexing::Continious => x,
             TilesetIndexing::Special(map) => map[&x],
         }
     }
